+++ Sources/TuistCoreTesting/Graph/TargetNode+TestData.swift
@@ -6,12 +6,10 @@ import TuistGraph
 public extension TargetNode {
     static func test(project: Project = .test(),
                      target: Target = .test(),
-                     dependencies: [GraphNode] = [],
-                     prune: Bool = false) -> TargetNode
+                     dependencies: [GraphNode] = []) -> TargetNode
     {
         TargetNode(project: project,
                    target: target,
-                   dependencies: dependencies,
-                   prune: prune)
+                   dependencies: dependencies)
     }
 }